implement generic download
